# apps/web/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MeadTools API",
    "version": "1.0.0",
    "description": "OpenAPI documentation for MeadTools public, authenticated user, and device integration endpoints."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current MeadTools host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Send the access token in the Authorization header as `Bearer <token>`. External API clients should use MeadTools access tokens returned by the auth endpoints."
      }
    },
    "schemas": {
...
      "AdminAuthErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Authorization header missing",
              "Token missing",
              "Invalid token or unauthorized access",
              "Invalid or expired token",
              "User not found",
              "Server misconfiguration",
              "Unauthorized access",
              "Forbidden – admin access required.",
              "Failed to verify admin"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuthenticatedRouteErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Authorization header missing",
              "Token missing",
              "Invalid token or unauthorized access",
              "Invalid or expired token",
              "User not found",
              "Server misconfiguration"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "BjcpIngredientsResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/BjcpIngredientResponse"
        }
      },
      "BjcpIngredientsFetchErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Failed to fetch ingredients"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateBjcpIngredientRequestBody": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "BjcpIngredientResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "created_at",
          "label",
          "category",
          "value"
        ]
      },
      "CreateBjcpIngredientFailureErrorResponse": {
        "$ref": "#/components/schemas/ApiErrorResponse"
      },
...
      "UpdateBrewEntryRequestBody": {
        "$ref": "#/components/schemas/brewEntryMutationObjectSchema"
      },
      "UpdateBrewEntryResponse": {
        "type": "object",
        "properties": {
          "entry": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BrewEntryWithBrewIdResponse"
              }
            ]
          }
        },
        "required": [
          "entry"
        ]
      },
      "BrewValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Invalid JSON body",
              "Missing recipe_id",
              "Missing brew_id",
              "Missing entry_id",
              "Invalid client_entry_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "BrewEntryUpdateErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Failed to update entry.",
              "Server misconfiguration"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "DeleteBrewEntrySuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "enum": [
              "Entry deleted successfully."
            ]
          }
        },
        "required": [
          "message"
        ]
      },
      "BrewEntryDeleteErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Failed to delete entry.",
              "Server misconfiguration"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "UpdateRecipeCommentRequestBody": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "comment"
        ]
      },
      "UpdateRecipeCommentResponse": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/components/schemas/RecipeCommentResponse"
          }
        },
        "required": [
          "comment"
        ]
      },
      "UpdateRecipeCommentValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Comment text is required"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "UpdateRecipeCommentFailureErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Failed to update comment",
              "Server misconfiguration"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "DeleteRecipeCommentResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "deleted_at": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "deleted_at"
            ]
          }
        },
        "required": [
          "deleted"
        ]
      },
      "DeleteRecipeCommentFailureErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Failed to delete comment",
              "Server misconfiguration"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "ContactRequestBody": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "user_name",
          "user_email",
          "message"
        ]
      },
      "ContactSuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "enum": [
              "Email sent successfully"
            ]
          }
        },
        "required": [
          "message"
        ]
      },
      "ContactValidationErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "enum": [
              "All fields are required"
            ]
          }
        },
        "required": [
          "message"
        ]
      },
      "ContactSendErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "enum": [
              "Failed to send email"
            ]
          }
        },
        "required": [
          "message"
        ]
      },
...
      "jsonObjectSchema": {
        "type": "object",
        "additionalProperties": {}
      },
...
      "brewEntryMutationObjectSchema": {
        "type": "object",
        "properties": {
          "datetime": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "gravity": {
            "type": "number",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "nullable": true
          },
          "temp_units": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TemperatureUnitResponse"
              }
            ]
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/jsonObjectSchema"
              }
            ]
          }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "BrewEntryTypeResponse": {
        "type": "string",
        "enum": [
          "NOTE",
          "GRAVITY",
          "VOLUME",
          "TEMPERATURE",
          "PH",
          "ADDITION",
          "NUTRIENT",
          "RACKING",
          "STABILIZATION",
          "BACKSWEETENING",
          "PACKAGING",
          "TASTING",
          "STAGE_CHANGE",
          "ISSUE"
        ]
      },
      "BrewEntryWithBrewIdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "datetime": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/BrewEntryTypeResponse"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "gravity": {
            "type": "number",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "nullable": true
          },
          "temp_units": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TemperatureUnitResponse"
              }
            ]
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/jsonObjectSchema"
              }
            ]
          },
          "user_id": {
            "type": "number",
            "nullable": true
          },
          "brew_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "datetime",
          "type",
          "title",
          "note",
          "gravity",
          "temperature",
          "temp_units",
          "data",
          "user_id",
          "brew_id"
        ]
      },
...
      "RecipeCommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "recipe_id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          },
          "parent_id": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "recipe_id",
          "user_id",
          "parent_id",
          "comment",
          "created_at",
          "updated_at"
        ]
      },
...
      "TemperatureUnitResponse": {
        "type": "string",
        "enum": [
          "F",
          "C",
          "K"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/bjcp-ingredients": {
      "get": {
        "operationId": "get-bjcp-ingredients",
        "summary": "List BJCP ingredients",
        "description": "Returns the BJCP ingredient catalog used by recipe-building tools.",
        "tags": [
          "BJCP Ingredients"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BjcpIngredientsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BjcpIngredientsFetchErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-bjcp-ingredients",
        "summary": "Create BJCP ingredient",
        "description": "Admin-only. Creates a BJCP ingredient catalog entry.",
        "tags": [
          "Admin"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBjcpIngredientRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BjcpIngredientResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBjcpIngredientFailureErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/brews/{brew_id}/entries/{entry_id}": {
      "patch": {
        "operationId": "patch-brews-{brew_id}-entries-{entry_id}",
        "summary": "Update brew entry",
        "description": "Updates a timeline entry for a brew. Stage change entries only allow datetime and note edits.",
        "tags": [
          "Brews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "brew_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          },
          {
            "in": "path",
            "name": "entry_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrewEntryRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBrewEntryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrewValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrewEntryUpdateErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-brews-{brew_id}-entries-{entry_id}",
        "summary": "Delete brew entry",
        "description": "Deletes a timeline entry for a brew. Stage change entries cannot be deleted through this endpoint.",
        "tags": [
          "Brews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "brew_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          },
          {
            "in": "path",
            "name": "entry_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteBrewEntrySuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrewValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrewEntryDeleteErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "operationId": "post-contact",
        "summary": "Send contact message",
        "description": "Sends a contact form message to the MeadTools support inbox.",
        "tags": [
          "Contact"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactValidationErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactSendErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/comments/{id}": {
      "patch": {
        "operationId": "patch-comments-{id}",
        "summary": "Update recipe comment",
        "description": "Updates one of the authenticated user's own recipe comments.",
        "tags": [
          "Recipes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecipeCommentRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRecipeCommentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRecipeCommentValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateRecipeCommentFailureErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-comments-{id}",
        "summary": "Delete recipe comment",
        "description": "Soft-deletes a recipe comment. Users can delete their own comments; admins can delete any comment.",
        "tags": [
          "Recipes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRecipeCommentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedRouteErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRecipeCommentFailureErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}